refactored collection context, collection page and collection container
