+++ src/app/collection/page.tsx
@@ -1,9 +1,17 @@
+import { getOne } from "@/lib/collection/getUserCollection";
 import CollectionContainer from "@/ui/CollectionContainer";
+import { getServerSession } from "next-auth";
 
-export default function Collection() {
+export default async function Collection() {
+  const session = await getServerSession();
+
+  if (!session?.user?.email) {
+    return <p>Connecte toi pour consulter ta collection</p>;
+  }
+  const collection = await getOne(session?.user?.email);
   return (
     <>
-      <CollectionContainer />
+      <CollectionContainer collection={collection} />
     </>
   );
 }
