+++ src/context/CollectionContext.tsx
@@ -1,19 +1,18 @@
 "use client";
 
+import { getOne } from "@/lib/collection/getUserCollection";
 import { CardsModel } from "@/model/CardsModel";
 import { createContext, useContext, useEffect, useState } from "react";
 import { useUserContext } from "./UserContext";
 
 type CollectionContextType = {
-  collection: CardsModel[] | [];
-  loading: boolean;
-  refreshCollection: () => Promise<void>;
+  collection: CardsModel[];
+  setCollection: React.Dispatch<React.SetStateAction<CardsModel[]>>;
 };
 
 const CollectionContext = createContext<CollectionContextType>({
   collection: [],
-  loading: true,
-  refreshCollection: async () => {},
+  setCollection: () => {},
 });
 
 export const useCollectionContext = () => useContext(CollectionContext);
@@ -25,32 +24,29 @@ export function CollectionProvider({
 }) {
   const { user } = useUserContext();
   const [collection, setCollection] = useState<CardsModel[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  const fetchCollection = async () => {
-    if (!user?.email) return;
-
-    setLoading(true);
-    try {
-      const res = await fetch(`/api/users/${user.email}/collections`);
-      const data: CardsModel[] = await res.json();
-      setCollection(data);
-    } catch (err) {
-      console.error("Erreur fetch collection context :", err);
-      setCollection([]);
-    } finally {
-      setLoading(false);
-    }
-  };
 
   useEffect(() => {
+    const fetchCollection = async () => {
+      if (!user?.email) return;
+
+      try {
+        const data = await getOne(user.email);
+        if (Array.isArray(data)) {
+          setCollection(data);
+        } else {
+          throw new Error("Format de données invalide");
+        }
+      } catch (err) {
+        console.error("Erreur fetch collection context :", err);
+        setCollection([]);
+      }
+    };
+
     fetchCollection();
   }, [user?.email]);
 
   return (
-    <CollectionContext.Provider
-      value={{ collection, loading, refreshCollection: fetchCollection }}
-    >
+    <CollectionContext.Provider value={{ collection, setCollection }}>
       {children}
     </CollectionContext.Provider>
   );
