+++ src/lib/collection/getUserCollection.ts
@@ -1,9 +1,9 @@
 import { collectionMessages } from "@/data/responseMessages";
-import { apiRoutes } from "@/data/ROUTES";
+import { apiUrl } from "@/data/ROUTES";
 import { CardsModel } from "@/model/CardsModel";
 
-export async function getOne(user_id: number): Promise<CardsModel[]> {
-  const res = await fetch(`${apiRoutes.COLLECTIONS}/${user_id}`, {
+export async function getOne(email: string): Promise<CardsModel[]> {
+  const res = await fetch(`${apiUrl}/api/users/${email}/collections`, {
     method: "GET",
     headers: {
       Accept: "application/json",
